AirspeedValidator: remove load_factor check (below stall speed check)
This check was triggering an airspeed failure if the measured airspeed
falls below the set stall speed. The assumption was that it's theoretically
impossibl to fly normally below stall, and thus it must be that the airspeed
measurement must be off. In reality, flying below stall speed happens often,
either at real stalls or becasue the stall parameter was set to high for the
vehicle. Not all stalls are fatal, acutally most systems can recover well from
one, and it's safer if the airspeed doesn't get invalidated when it happens.
Airspeed sensor clogging can be detected by the innovation check, so I also
don't see a clear need to keep this check for that usecase.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>